[FIX] : AsyncConfig 설정
