--- src/main/java/Lumo/lumo_backend/domain/alarm/entity/config/AsyncConfig.java
@@ -1,24 +1,0 @@
-package Lumo.lumo_backend.global.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@Configuration
-@EnableAsync
-public class AsyncConfig {
-
-    @Bean(name = "alarmTaskExecutor")
-    public Executor alarmTaskExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(2);
-        executor.setMaxPoolSize(5);
-        executor.setQueueCapacity(50);
-        executor.setThreadNamePrefix("alarm-async-");
-        executor.initialize();
-        return executor;
-    }
-}
